Update checklists docs
